Fixed up some small things when studying
Mostly \left( and \right) replacing \big( and \big)
when the sizing looks a bit off.
Also I switched to centered dots (\cdots) when doing something like
1 + 2 + 3 + ...
since that is more standard.
There was also one missing right paren in there I think,
or at least a \big( with a normal ).